+++ R/grid_nested.R
@@ -3,6 +3,8 @@
 #' Creates a panel for a layout with its own internal gridlayout
 #'
 #' @inheritParams grid_card
+#' @param title Optional title for the card. Gets wrapped in
+#'   `bslib::card_header()`
 #' @inheritParams grid_container
 #'
 #' @seealso [grid_card], [grid_container]
@@ -41,29 +43,4 @@ grid_nested <- function(area,
       nested_grid
     )
   )
-
-  # grid_place(
-  #   area = area,
-  #   htmltools::div(
-  #     class = "card no-pad",
-  #     if (!is.null(title)) {
-  #       card_header_internal(title)
-  #     },
-  #     htmltools::tags$div(
-  #       class = "card-body",
-  #       nested_grid
-  #     )
-  #   )
-  # )
-
-  # grid_card(
-  #   area = area,
-  #   title = title,
-  #   htmltools::tags$div(
-  #     class = "card-body",
-  #     nested_grid
-  #   )
-  # )
 }
-
-