+++ man/examples/simple_app_old.R
@@ -12,12 +12,10 @@ shinyApp(
     row_sizes = c("50px", "1fr"),
     col_sizes = c("200px", "1fr"),
     grid_card_text("header", "This is my header"),
-    grid_card(
-      area = "sidebar",
-      card_header("Settings"),
-      card_body(
-        sliderInput("bins", "Number of bins:", 1, 50, 30, width = "100%")
-      )
+    grid_card_old(
+      "sidebar",
+      title = "Settings",
+      sliderInput("bins","Number of bins:", 1, 50, 30, width = "100%")
     ),
     grid_card_plot("distPlot")
   ),