+++ NEWS.md
@@ -1,8 +1,10 @@
-# gridlayout 0.1.1
+# gridlayout 0.2.0
 
-### Major new features and improvements
+### Major changes
 
 - `grid_card()` now wraps `bslib::card()` directly. This means that there are some api changes such as depreciated the `title` argument for using `bslib::card_header(title)` instead. Old `grid_card()` arguments that are present in code will get flagged with a warning to allow users to know when/how to update their code with the new version.
+- `card_plot_output()` is now depreciated in favor of the new `bslib` cards smart fill behavior. Simple plots can still be done with `grid_card_plot()`.
+- The function `grid_card_old()` has been added to ease transition to the new card api and to keep some features such as collapsibility are not available on the new api.
 
 # gridlayout 0.1.0